+++ lib/flutter_themez.dart
@@ -1,6 +1,7 @@
 library flutter_themez;
 
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 
 part 'src/extensions.dart';
 part 'src/flutter_themez.dart';